Feature/jijc new

See merge request !53

+++ src/main/java/com/meiren/interceptor/LoginInterceptor.java
@@ -68,7 +68,7 @@ public class LoginInterceptor implements HandlerInterceptor {
                 BeanUtils.copyProperties(user, userVO);
                 //当老的session没有失效且用户信息和sso校验得到的用户信息一致时，直接用老的session
                 //避免uuid每次sso校验后都重置
-                if(oldUserVO != null && (oldUserVO.getId() == userVO.getId())){
+                if(oldUserVO != null && (oldUserVO.getId().longValue() == userVO.getId().longValue())){
                     return true;
                 }
                 userVO.setInSide(this.isInSide(user.getId()));
